+++ backend/grants/admin/__init__.py
@@ -1,0 +1,1 @@
+from .grant import GrantAdmin  #  noqa
